fix: Org admin cannot use system tools.
